modules/locations_kafka_consumer/requirements.txt
GeoAlchemy2==0.8.4
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
Shapely==1.7.0
SQLAlchemy==1.3.19
kafka-python==2.0.2
psycopg2-binary==2.9.6